refactor: restructure SignUp and RecruitmentHeader components
